validations.js added
