+++ back-end/controllers/snackController.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const confirmHealth = require('../confirmHealth.js');
-const correctSnack = require('../spellCheck');
+const { correctSnack, checkName , isBoolean } = require('../validations');
 const snacks = express.Router();
 
 const {getAllSnacks, getOneSnack, createSnack , deleteSnack, updateSnack} = require("../queries/snacks.js");
@@ -24,7 +24,7 @@ snacks.get("/:id", async (req,res)=>{
     try{
         const snack = await getOneSnack(id);
         if(snack.id){
-            res.status(200).json({success: true, payload: snack, });
+            res.status(200).json({success: true, payload: snack});
         } else {
             res.status(404).json({success: false, payload: "Snack not found"});
         }
